Separate MaatAPI and service models
