+++ crime-hardship/src/main/java/uk/gov/justice/laa/crime/hardship/dto/ErrorDTO.java
@@ -6,6 +6,6 @@ import lombok.Value;
 @Value
 @Builder
 public class ErrorDTO {
-    private String code;
-    private String message;
+    String code;
+    String message;
 }
