+++ crime-hardship/src/main/java/uk/gov/justice/laa/crime/hardship/dto/HardshipReviewDTO.java
@@ -1,15 +1,17 @@
 package uk.gov.justice.laa.crime.hardship.dto;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import uk.gov.justice.laa.crime.hardship.dto.maat_api.SolicitorCosts;
+import uk.gov.justice.laa.crime.hardship.model.*;
+import uk.gov.justice.laa.crime.hardship.staticdata.enums.CourtType;
+import uk.gov.justice.laa.crime.hardship.staticdata.enums.HardshipReviewResult;
 import uk.gov.justice.laa.crime.hardship.staticdata.enums.HardshipReviewStatus;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 @Data
@@ -20,41 +22,21 @@ public class HardshipReviewDTO {
 
     private Integer id;
     private Integer cmuId;
+    private CourtType courtType;
+    private HardshipReviewStatus reviewStatus;
+    private LocalDateTime reviewDate;
+    private ApiUserSession userSession;
+    private NewWorkReason reviewReason;
     private String notes;
     private String decisionNotes;
-    private LocalDateTime reviewDate;
-    private String reviewResult;
-    private BigDecimal disposableIncome;
-    private BigDecimal disposableIncomeAfterHardship;
-    private NewWorkReason newWorkReason;
-    private SolicitorCosts solicitorCosts;
-    private HardshipReviewStatus status;
-
-    @JsonIgnore
-    private LocalDateTime dateCreated;
-    @JsonIgnore
-    private Integer repId;
-    @JsonIgnore
-    private String userCreated;
-    @JsonIgnore
-    private LocalDateTime updated;
-    @JsonIgnore
-    private String userModified;
-    @JsonIgnore
-    private LocalDateTime resultDate;
-    @JsonIgnore
-    private String courtType;
-    @JsonIgnore
     private Integer financialAssessmentId;
-    @JsonIgnore
-    private String valid;
+    private BigDecimal totalAnnualDisposableIncome;
+    private SolicitorCosts solicitorCosts;
+    private List<OtherFundingSources> otherFundingSources;
+    private List<ExtraExpenditure> extraExpenditure;
+    private List<DeniedIncome> deniedIncome;
+    private List<HardshipProgress> progressItems;
+    private HardshipReviewResult reviewResult;
+    private BigDecimal postHardshipDisposableIncome;
 
-    @Builder.Default
-    private List<HardshipReviewDetail> reviewDetails = new ArrayList<>();
-    @Builder.Default
-    private List<HardshipReviewProgress> reviewProgressItems = new ArrayList<>();
-
-    public LocalDateTime getTimestamp() {
-        return updated != null ? updated : dateCreated;
-    }
 }
