+++ crime-hardship/src/main/java/uk/gov/justice/laa/crime/hardship/staticdata/enums/RequestType.java
@@ -1,0 +1,5 @@
+package uk.gov.justice.laa.crime.hardship.staticdata.enums;
+
+public enum RequestType {
+    CREATE, UPDATE
+}
