+++ crime-hardship/src/test/java/uk/gov/justice/laa/crime/hardship/service/MaatCourtDataServiceTest.java
@@ -10,7 +10,7 @@ import uk.gov.justice.laa.crime.commons.client.RestAPIClient;
 import uk.gov.justice.laa.crime.hardship.config.MockServicesConfiguration;
 import uk.gov.justice.laa.crime.hardship.config.ServicesConfiguration;
 import uk.gov.justice.laa.crime.hardship.data.builder.TestModelDataBuilder;
-import uk.gov.justice.laa.crime.hardship.dto.HardshipReviewDetail;
+import uk.gov.justice.laa.crime.hardship.dto.maat_api.HardshipReviewDetail;
 
 import java.util.List;
